requirements.txt
# API сервер
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2

# База данных
sqlalchemy==2.0.35

# Аналитика (Этапы 4-6)
pandas==2.2.3
scipy==1.14.1

# AI слой (Этап 8)
anthropic==0.40.0

# Dashboard (Этап 7)
streamlit==1.39.0

# HTTP клиент (симулятор)
httpx==0.27.2

# Утилиты
python-dotenv==1.0.1
